% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_frags.R
\name{filter_frags}
\alias{filter_frags}
\title{Filter Digested Fragments}
\usage{
filter_frags(frags, min, max)
}
\arguments{
\item{frags}{input frags data frame; output of digest_genome()}

\item{min}{minimum length of fragments to be included. Shorter fragments will be discarded}

\item{max}{maximum length of fragments to be included. Larger fragments will be discarded}
}
\value{
Returns a filtered frags data frame
}
\description{
Simply filters digested fragments based on user input paramaters
}
