% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sites.R
\name{get_sites}
\alias{get_sites}
\title{Get Sites}
\usage{
get_sites(genome, digest, sites)
}
\arguments{
\item{genome}{String indicating the species to be loaded, e.g. 'human' will
load the package 'BSgenome.Hsapiens.UCSC.hg38'}

\item{digest}{Object containing the output of get_digest()}

\item{sites}{Sequence to be scanned for (usually sites that can be probed
for chromatin accessibility); can take more than one entry;
e.g. sites = c('HCG', 'GCH')}
}
\value{
Returns data frame object with chromosome name, start position,
end position, length of fragment, and one column for each sites with the
number of sites per fragment
}
\description{
Finds the location and number of sites of interest
}
