% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neb_table_doc.R
\docType{data}
\name{neb.table}
\alias{neb.table}
\title{Title}
\format{
An object of class \code{data.frame} with 207 rows and 7 columns.
}
\usage{
data(neb.table)
}
\description{
Data from NEB website for restriction enzymes........
}
\examples{
data(neb.table)
}
\keyword{datasets}
