% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_overlaps.R
\name{process_overlaps}
\alias{process_overlaps}
\title{Process Overlaps}
\usage{
process_overlaps(frags.filt)
}
\arguments{
\item{frags.filt}{A filtered frags object}
}
\value{
Returns data frame object with number of fragments and
percent feature coverage for each genomic feature
}
\description{
Calculates the overlap of digested fragments with genomic features
}
